encoding/base32: decoder output depends on chunking of underlying reader
After an analysis, I figured that a way to do it could be to check, after
the call to readEncodedData whether the decoder already saw the end or not.

Yet, if we try to decode `I4======N4======`, the error message is `illegal base32 data at input byte 2`.
Therefore, if we split the input into two chunks `I4======` and `N4======` we want the error message to be the same and highlight that the error is at index 2.
This is why I added as part of the `decoder` struct a field representing the index of the latest non-padding character and made `decode` return the number of non-padding characters.

Fixes https://github.com/golang/go/issues/38657

Change-Id: I06fd718ea4ee6ded2cb26c2866b28581ad86e271
GitHub-Last-Rev: fb54ba35da4a4404fe790d4956f04c9afcaa008d
GitHub-Pull-Request: golang/go#52631