runtime/metrics: add /gc/cpu/limiter-overflow:cpu-seconds metric
This change exports the monotonically-increasing CPU-time overflow of GC
CPU time spent beyond the limiter's threshold. Since this overflow
directly corresponds to the enabling of the limiter, any change of this
metric directly indicates that the GC is actively being limited at the
expense of memory, and may be useful in deducing the cause of an
out-of-memory condition.

For #48409.

Change-Id: Ic6383b19e88058366a74f6ede1683b8ffb30a69c