time/tzdata: encode the zipfile as a single giant string
The contents of the string are not meant for human consumption, and
the repeated + operators for each line make it tedious for tools to
parse.

As a single line, the file is approximately equally deadly for IDEs
and faster for other tools (such as cmd/gofmt and cmd/compile).

For #43350.

Change-Id: I6dba4ec490038425c2ec44bd4fa5787f35079fb2