cmd{cover,covdata,go}: better coverage for tests that build tools
Some of the unit tests in Go's "cmd" tree wind up building a separate
copy of the tool being tested, then exercise the freshly built tool as
a way of doing regression tests. The intent is to make sure that "go
test" is testing the current state of the source code, as opposed to
whatever happened to be current when "go install <tool>" was last run.

Doing things this way is unfortunately rather unfriendly for coverage
testing. If I run "go test -cover cmd/mumble", and the cmd/mumble test
harness builds a fresh copy of mumble.exe, any runs of that new
executable won't generate coverage data.

As a proof of concept, this patch updates the test harness builder in
"cmd/cover" so that if "go test -cover" is in effect, we build the
tool executable for coverage as well. Doing this brings up the overall
test coverage number for cmd/cover quite dramatically:

before change:

  $ go test -cover .
  ok  	cmd/cover	1.100s	coverage: 1.5% of statements

after change:

  $ go test -cover .
  ok  	cmd/cover	1.299s	coverage: 84.2% of statements

Getting this to work requires a small change in the Go command as
well, to set GOCOVERDIR prior to executing a test binary.

This is not a perfect or comprehensive solution; many more patches
like this will be needed to address specific cases within the Go "cmd"
tree.

Updates #51430.

Change-Id: Ifcf0ea85773b80fcda794aae3702403ec8e0b733