cmd/compile: fix interface print in debugger for dynamic link
runtime._type, runtime.itab and the subtype of them were not added to builtin.
It's ok for static link, because the real type info can be dump in runtime package.
But for dynamic link, we can't use a fake type anymore,
all the fields of runtime.iface and runtime.efac must be real type.

For #52209

Change-Id: I2fdfb8eb9a2e80d6aebea5be47842f75f2537b76