os, syscall: don't consider stderr output as part of hostname
A successful invocation of the hostname command prints the hostname
to stdout and exits with code 0. No part of the hostname is printed
to stderr, so don't consider it.

This avoids false positive failures in environments where hostname
prints some extraneous information (such as performance warnings)
to stderr, and makes the test a bit more robust.

Fixes golang/go#52781.

Change-Id: I46aa6fbf95b6616bacf9c2b5e412b0851b230744