spec: use original terminology in metasyntax (cleanup)
The metasyntax used in the spec is exactly the Wirth Syntax
Notation (WSN), which eventually influenced EBNF. Be honest about
it and use the original terms in the productions. Specifically,
use the words "Term" and "Factor" rather than "Alternative"
and "Term". With this change the productions match WSN productions
exactly. It also resolves an inconsistency in the subsequent prose
which is referring to the correct "terms" now.

While at it, add a production for the entire Syntax itself,
again matching the original WSN definitions.

Also, replace the two uses of "grammar" with "syntax" for
consistency ("syntax" is the prevalent term used throughout
the spec).

Fixes #50074.

Change-Id: If770d5f32f56f509f85893782c1dafbb0eb29b2e