cmd/compile: strip counter instrumentation from some hot runtime funcs
DO NOT SUBMIT

This patch introduces an experimental pass that strips out coverage
counter updates when compiling specific selected hot runtime
functions, so as to avoid slowdowns due to contention for counter vars
on systems with high core counts.

Not ready for prime time (also not really clear whether this is the
best way to handle this situation in general).

Change-Id: I2c68eaa3139d134f89dbbaa98406161e32d6dbd4