cmd/compile/internal/noder: remove another unified IR assumption about types.LocalPkg.Path
The go/importer.TestForCompiler/LookupCustom test tries to read in the
export data for "math/big", but with a package path of "math/bigger"
instead. This has historically worked because the export data formats
were designed to not assume the package's own path, but I expect we
can safely remove support for this now.

However, since that would be a user-visible change, for now just tweak
unified IR to not break that assumption so we can land CL 393715. We
can revisit whether it's actually safe to break that go/importer use
case later.

Updates #51734.

Change-Id: I5e89314511bd1352a9f5e14a2e218a5ab00cab3c