cmd/link: call syscall.FlushFileBuffers on outbuf Unmap
DO NOT SUBMIT

In the windows version of OutBuf.munmap, call syscall.FlushFileBuffers
after the call to syscall.FlushViewOfFile, on the theory that this
will help flush all associated meta-data for the file the linker is
writing.

Note: at the moment doesn't pass bootstrap on windows, failing with

  C:\workdir\go\pkg\tool\windows_amd64\link: FlushFileBuffers failed: The handle is invalid.

Not sure what the issue is.

Change-Id: Ibff7d05008a91eeed7634d2760153851e15e1c18