cmd/cover: restructure unit tests to eliminate use of TestMain
Rewrite the unit tests for cmd/cover to avoid the use of TestMain,
since it is not absolutely needed here. New setup has a single test
function TestCover that performs setup and the forks off a bunch of
sub-tests. No change in test functionality. This refactoring is
intended to make subsequent patches easier to write.

Change-Id: Idc40ce5ac727e3c236bbc04f78c7b4f25e473599