runtime: eliminate cache miss in findObject path
A whopping 50% of cache misses incurred by the GC while scanning come
from the first access to the mspan data structure in findObject.

This change eliminates this cache miss by inlining oft-used fields of
the mspan into the span map. This change quadruples the memory overhead
of the spans array in the heapArena structure, but this only accounts
for an increase of 24/8192 = 0.3% increase in overall memory usage.
However, I expect a much, much larger increase in GC efficiency as a
result.

Change-Id: I8bce3d25e05e61a77027b131c9101c4c5fd44bcf