runtime: allow page marks to be set on any page of the span
DO NOT SUBMIT

Needs benchmarks.

This change eliminates a couple of cache misses from greyobject by
skipping a span lookup for the page mark bits, and instead reusing the
heap arena pointer discovered by findObject. Because this heap arena may
not contain the first page of the span we're marking, the page reclaimer
needs to be robust to this case.

As long as we continue to manage the pageInUse bitmap, this is
straightforward: upon encountering an in-use span bit, we begin tracking
to see if we find any marked pages before either the next in-use span
bit, or the end of the chunk. If we see any marked pages before then, we
know the span is marked. Otherwise, we try to reclaim it.

This makes the page reclaimer a bit slower than before, but greyobject
is significantly hotter than the page reclaimer so overall the tradeoff
is worth it.

Change-Id: Ia1573d4eadacbd1c45e7ef1db6a5f15a79eac7a2