runtime: remove heap lock acquisition in the page reclaimer
The only remaining reason why the page reclaimer needs to lock the heap
is because it blocks the freeing of spans.

If a span is freed concurrently with the page reclaimer executing, it
could observe a stale span pointer because of a concurrent update to
a pageInUse bit. If that stale span has already been freed and then
reallocated for another purpose, the page reclaimer might still try to
sweep it (erroneously). (The comment in the page reclaimer is somewhat
stale as it discusses freeing and merging spans, which is an
implementation detail from many releases ago. The core issue of freeing
is still there, though.)

Resolve this potential race by checking the span state. If such a
concurrent free happens and the page reclaimer observes a stale span
pointer, then there are three cases:

1) The span is free and dead, so it may be safely skipped.
2) The span has been reused for a non-heap purpose. It may be safely
   skipped.
3) The span has been reused for the heap. In this case, the span's state
   is published last, so if we observe the span's state as a heap span,
   then we can safely assume its sweepgen is set correctly.
   For a reused span like this, the page reclaimer will fail to acquire
   it because its sweepgen is for the current cycle.

Change-Id: Ibc7d5ea42732f5564bdb9d02bf727deeaf503609