strings: add CommonPrefixLen function
The CommonPrefixLen function returns the length of the common
prefix of two strings. Although the naive implementation is very
simple (3 LOC), the need for this algorithm appears in many domains
(e.g. diff algorithms) and it's possible to significantly outperform
the naive implementation by, for example, comparing word-sized
chunks, as this implementation does, for a 5x asymptotic gain.

Opportunities for further significant optimizations are
noted in the comment.

The bytes package may want a similar function.

See proposal https://github.com/golang/go/issues/53120

Change-Id: I142c1334e35d81c611fddc37feca6389e19454ec