runtime: don't inspect the stack for delayed signals from TSAN
TSAN intercepts signals and calls the signal handler at a later
time. When the signal handler is called, the memory may have
changed, but the signal context remains old. As the signal context
and the memory don't match, it is unsafe to unwind the stack from
the signal PC and SP. We have to ignore the signal.

It is probably als not safe to do async preemption, which relies
on the signal PC, and inspects and even writes to the stack (for
call injection).

We also inspect the stack for fatal signals (e.g. SIGSEGV), but I
think they are not delayed. For other signals we don't inspect
the stack, so they are probably fine.

Fixes #27540.

Change-Id: I5c80a7512265b8ea4a91422954dbff32c6c3a0d0