cmd/compile, go/types: make validType0 walk/check more accurate
The old version maintained an environment and a map from type to
valid/invalid/unknown/marked, and would pop from the environment
when translating TypeParams, bit did not reset the marked status
for corrsponding types as the environment was popped

The new version retains the cheap test for unknown/marked, but
also maintains an accurate set of really-marked types (implemented
as a slice).  The cheap test is conservative, but in the case that
it reports a false positive (i.e., a type cycle error) the accurate
set can be double-checked.  To ensure that types are compared
accurately, Named types are canonicalized.

This is an interim version of the fix; a proposed further change
is to do away with the conservative version of the test
and only use the accurate set.

Fixes #52698.

Change-Id: I03d70e9039826e48ba37070ce2309f8e5e47582c