cmd/compile, go/types: make validType0 walk/check more accurate
The old version maintained an environment and a path stack,
and would pop from the environment when translating TypeParams,
but did not also pop from the corresponding path stack.

The old version also used a simple unknown/marked/valid/invalid
bit to keep detect cycles, and this was wrong because it did not
model the path-popping added above.

The new version retains the cheap test for unknown/marked, but
double checks that the "marked" type actually appears on the
stack, and treats a type as "unknown" if it does not.

Fixes #52698.

Change-Id: I03d70e9039826e48ba37070ce2309f8e5e47582c