reflect: allow Value be stack allocated
Currently, reflect.ValueOf forces the referenced object to be heap
allocated. This CL makes it possible to be stack allocated. We
need to be careful to make sure the compiler's escape analysis can
do the right thing, e.g. channel send, map assignment, unsafe
pointer conversions.

Tests will be added in a later CL.

TODO: how do we ensure the correctness? CL 408827 might help, but
not sure if it is good enough.

Change-Id: I8663651370c7c8108584902235062dd2b3f65954