runtime: only use CPU time from the current window in the GC CPU limiter
Currently the GC CPU limiter consumes CPU time from a few pools, but
because the events that flush to those pools may overlap, rather than be
strictly contained within, the update window for the GC CPU limiter, the
limiter's accounting is ultimately sloppy.

This sloppiness complicates accounting for idle time more completely,
and makes reasoning about the transient behavior of the GC CPU limiter
much more difficult.

To remedy this, this CL adds a field to the P struct that tracks the
start time of any in-flight event the limiter might care about, along
with information about the nature of that event. This timestamp is
managed atomically so that the GC CPU limiter can come in and perform a
read of the partial CPU time consumed by a given event. The limiter also
updates the timestamp so that only what's left over is flushed by the
event itself when it completes.

The end result of this change is that, since the GC CPU limiter is aware
of all past completed events, and all in-flight events, it can much more
accurately collect the CPU time of events since the last update. There's
still the possibility for skew, but any leftover time will be captured
in the following update, and the magnitude of this leftover time is
effectively bounded by the update period of the GC CPU limiter, which is
much easier to consider.

One caveat of managing this timestamp-type combo atomically is that they
need to be packed in 64 bits. So, this CL gives up the top 8 bits of the
timestamp and places the type information there. What this means is we
effectively have only a 56-bit resolution timestamp. This is fine when
the application is active, because we only use this timestamp to compute
durations, and the durations are bounded by the update period of the
limiter. When actually computing a difference, we assume the top 8 bits
of the start timestamp are the same as the current timestamp.

However, 56 bits of resolution (or, 2^56 nanosecond = 2.3 year
durations) might not be enough if the application isn't actively
executing. Even this case is OK. To understand why, let's take a
concrete example: suppose someone's laptop is running a Go daemon and
the laptop is closed but the machine clock continues marching forward.
After at least 2.3 years the laptop is opened and the daemon begins
executing again. In this case, by assuming the top 8 bits of the old
timestamp are the same as the new, the Go program might observe that the
old timestamp exceeds the new. I don't think it's reasonable to crash in
this case; if we simply ignore what happened, the worst that can happen
is some transient bad behavior in the GC, but it's very unlikely for
that bad behavior to lead to an OOM (unless it was already in thrashing
territory when the laptop first closed). After a few seconds of
execution, the GC CPU limiter should be back to normal.

If that scenario sounds absurdly specific, then I think I demonstrated
my point that it's really not worth worrying about.

For #52890.

Change-Id: I347f30ac9e2ba6061806c21dfe0193ef2ab3bbe9